devel/requirements.txt
Django==2.2.9
pytz
mysqlclient
dateutils
IP2Location
django_redis
django-globals
easy-cache
easy-thumbnails
polib
django-ipware
djangorestframework
transliterate
requests
requests[security]
httplib2
django-render-block
django-templated-email
html2text
py3DNS
geopy
phonenumbers
cryptography
uwsgi
uwsgi_tasks
jsonfield
future
django-crispy-forms
django-formtools
django-import-export
pycryptodome>=3.7.0
python-telegram
